+++ components/CustomSwitch.tsx
@@ -15,7 +15,7 @@ const CustomSwitch = ({
             <Switch
                 trackColor={{ false: Colors.navy.p, true: Colors.navy.p }}
                 thumbColor={Colors.gray[5]}
-                ios_backgroundColor={Colors.gray[5]}
+                ios_backgroundColor={Colors.gray[3]}
                 onValueChange={toggleSwitch}
                 value={isEnabled}
             />
